chore: 기존 도메인 테스트 패키지들을 unit 하위로 이동 (#217)
