+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/PickGitApplicationTests.java
@@ -1,6 +1,6 @@
 package com.woowacourse.pickgit;
 
-import com.woowacourse.pickgit.post.PostTestConfiguration;
+import com.woowacourse.pickgit.unit.post.PostTestConfiguration;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
